* Merge from r7874-r7888. * Switch to 3.9.1 version.
git-svn-id: svn://localhost/gambas/branches/3.9@7889 867c0c6c-44f3-4631-809d-bfa615b0a4ec

+++ gb.db.odbc/src/main.c
@@ -134,7 +134,7 @@ void reportODBCError(char *fn,
 			ret = SQLGetDiagRecW(type, handle, ++i, state, &native, text, sizeof(text), &len);
 
 			if (SQL_SUCCEEDED(ret))
-				fprintf(stderr, "gb.db.odbc: %s:%d:%d:%s\n", (char *)state, i, native, (char *)text);
+				fprintf(stderr, "gb.db.odbc: %s:%d:%d:%s\n", (char *)state, (int)i, (int)native, (char *)text);
 		}
 		while (ret == SQL_SUCCESS);
 	}
